+++ CHANGELOG.rst
@@ -43,6 +43,7 @@ Changed
 - API requests are not longer logged to ``rasa_core.log`` by default in order to avoid
   problems when running on OpenShift (use ``--log-file rasa_core.log`` to retain the
   old behavior)
+- ``metadata`` attribute added to ``UserMessage``
 
 Fixed
 -----